Refactor util organization
